Stabilize gateway ui (#596)
* chore: refactor gateway Form components

chore: fix

* chore: bump fedimint version

* chore: disable unfinished feature - gateway onchain wallet

* fix(gateway): update gateway rpcs to match current version of fedimint

* chore: update translate script and run translations
