+++ apps/router/src/api/GatewayApi.ts
@@ -13,6 +13,7 @@ import {
   SpendEcashResponse,
   ReceiveEcashResponse,
   GatewayBalances,
+  WithdrawResponse,
 } from '@fedimint/types';
 import { GatewayConfig } from '../types/gateway';
 
@@ -129,6 +130,26 @@ export class GatewayApi {
     }
   };
 
+  // TODO: Implement gateway onchain wallet in UI
+  // fetchGatewayOnchainAddress = async (): Promise<string> => {
+  //   try {
+  //     const res: Response = await this.post('get_ln_onchain_address', {});
+
+  //     if (res.ok) {
+  //       const address: string = await res.json();
+  //       return Promise.resolve(address);
+  //     }
+
+  //     throw responseToError(res);
+  //   } catch (error) {
+  //     console.error('Error fetching gateway onchain address', error);
+  //     return Promise.reject({
+  //       message: 'Error fetching gateway onchain address',
+  //       error,
+  //     });
+  //   }
+  // };
+
   // TODO: unimplemented in fedimint
   backup = async (federationId: string): Promise<void> => {
     try {
@@ -210,7 +231,10 @@ export class GatewayApi {
     payload: CreateBolt11InvoiceV2Payload
   ): Promise<string> => {
     try {
-      const res: Response = await this.post('create_bolt11_invoice', payload);
+      const res: Response = await this.post(
+        'create_bolt11_invoice_for_operator',
+        payload
+      );
 
       if (res.ok) {
         const invoice: string = await res.json();
@@ -433,7 +457,7 @@ export class GatewayApi {
     }
   };
 
-  submitPegOut = async (payload: PegOutPayload): Promise<string> => {
+  submitPegOut = async (payload: PegOutPayload): Promise<WithdrawResponse> => {
     try {
       const res: Response = await this.post('withdraw', {
         federation_id: payload.federationId,
@@ -442,8 +466,8 @@ export class GatewayApi {
       });
 
       if (res.ok) {
-        const txid: string = await res.json();
-        return Promise.resolve(txid);
+        const response: WithdrawResponse = await res.json();
+        return Promise.resolve(response);
       }
 
       throw responseToError(res);
