+++ apps/router/src/gateway-ui/components/walletCard/OnchainCard.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useCallback } from 'react';
 import { Box, Button, Flex, Text } from '@chakra-ui/react';
 import { useTranslation } from '@fedimint/utils';
 import { FaArrowDown, FaArrowUp } from 'react-icons/fa';
@@ -13,21 +13,24 @@ export const OnchainCard = React.memo(function OnchainCard(): JSX.Element {
   const { t } = useTranslation();
   const { state, dispatch } = useGatewayContext();
 
-  const handleModalOpen = (action: WalletModalAction) => {
-    const federations = state.gatewayInfo?.federations;
-    if (!federations?.length) return;
+  const handleModalOpen = useCallback(
+    (action: WalletModalAction) => {
+      const federations = state.gatewayInfo?.federations;
+      if (!federations?.length) return;
 
-    dispatch({
-      type: GATEWAY_APP_ACTION_TYPE.SET_WALLET_MODAL_STATE,
-      payload: {
-        action,
-        type: WalletModalType.Onchain,
-        selectedFederation: federations[0],
-        showSelector: true,
-        isOpen: true,
-      },
-    });
-  };
+      dispatch({
+        type: GATEWAY_APP_ACTION_TYPE.SET_WALLET_MODAL_STATE,
+        payload: {
+          action,
+          type: WalletModalType.Onchain,
+          selectedFederation: federations[0],
+          showSelector: true,
+          isOpen: true,
+        },
+      });
+    },
+    [dispatch, state.gatewayInfo?.federations]
+  );
 
   return (
     <Box p={6} bg='white' borderRadius='lg' borderWidth='1px'>
@@ -62,8 +65,8 @@ export const OnchainCard = React.memo(function OnchainCard(): JSX.Element {
         <Text color='gray.600'>{t('wallet.onchain-description')}</Text>
       </Box>
 
-      {/* Node's Onchain Wallet Section */}
-      <Box>
+      {/* TODO: Implement Node's Onchain Wallet Section */}
+      {/* <Box>
         <Flex justify='space-between' align='center' mb={4}>
           <Text fontSize='md' fontWeight='medium'>
             {t('wallet.gateway-on-chain-wallet')}
@@ -82,7 +85,7 @@ export const OnchainCard = React.memo(function OnchainCard(): JSX.Element {
           </Flex>
         </Flex>
         <Text color='gray.600'>{t('wallet.gateway-onchain-manage')}</Text>
-      </Box>
+      </Box> */}
     </Box>
   );
 });
