+++ apps/router/src/languages/de.json
@@ -32,7 +32,8 @@
     "connect": "Verbinden",
     "password": "Passwort",
     "federation-name": "Bundesname",
-    "guardian-name": "Name des Vormunds"
+    "guardian-name": "Name des Vormunds",
+    "guardians": "Wächter"
   },
   "notConfigured": {
     "title": "Nicht konfiguriert",
@@ -417,7 +418,16 @@
     "to-address": "An Adresse",
     "txid": "Transaktions-ID",
     "claimed-note": "Anspruchsvolle Notiz",
-    "claimed-amount": "Geltend gemachter Betrag"
+    "claimed-amount": "Geltend gemachter Betrag",
+    "balances-summary": "Saldoübersicht",
+    "lightning-payments": "Blitzzahlungen",
+    "lightning-payments-description": "Zahlungen über das Lightning Network über die Föderation senden und empfangen",
+    "onchain-description": "Bitcoin in die Föderation einzahlen und abheben",
+    "gateway-onchain-manage": "Verwalten Sie Bitcoin in der Onchain-Brieftasche des Gateways.",
+    "federation-balances": "Föderationssalden",
+    "peg-in": "Peg-In",
+    "peg-out": "Peg-Außentritt",
+    "gateway-on-chain-wallet": "Gateway On-Chain Wallet"
   },
   "wallet-modal": {
     "title": "Wallet-Aktionen",
@@ -443,7 +453,11 @@
       "peg-out-success": "Peg Out TX gesendet!",
       "create-ecash": "Erstelle Ecash Note",
       "ecash-created": "Haben eine {{amount}} ecash-Note von Ihrem {{federationName}}-Guthaben abgehoben. Senden Sie es an den Empfänger, um es einzufordern.",
-      "ecash-error": "Fehler beim Erstellen der ecash-Notiz: {{error}}"
+      "ecash-error": "Fehler beim Erstellen der ecash-Notiz: {{error}}",
+      "lightning-invoice": "Blitzrechnung",
+      "lightning-invoice-placeholder": "lnbc...",
+      "pay-invoice": "Zahle Lightning-Rechnung",
+      "peg-out-error": "Fehler beim Auslagern auf der Blockchain"
     }
   },
   "federation-card": {
@@ -572,5 +586,14 @@
     "sentence-one": "Zahle an die Adresse oder scanne den QR-Code",
     "transaction-id": "Transaktions-ID",
     "create-pegin-address": "Peg-in-Adresse erstellen"
+  },
+  "router": {
+    "title": "Noch keine Dienste verbunden.",
+    "services-title": "Eine Fedimint-Föderation besteht aus zwei Arten von Dienstleistungen:",
+    "services-description": "Verantwortlich für die Durchführung des Fedimint-Protokolls, die Verwahrung von Mitteln und die Verwaltung der Prägung und Einlösung von eCash-Noten. Sie nutzen verteilten Konsens, um die Föderation abzusichern.",
+    "gateways": "Gateways (Lightning-Dienstanbieter):",
+    "gateways-description": "Brücke zwischen dem Fedimint und dem Lightning Network, die es Benutzern ermöglicht, Lightning-Zahlungen zu senden und zu empfangen. Sie können von der Föderation oder unabhängigen Anbietern betrieben werden.",
+    "learn-more": "Sie können hier mehr darüber erfahren, wie Sie Fedimint-Dienste einrichten.",
+    "learn-more-link": "Erfahren Sie, wie Sie Fedimint-Dienste einrichten können."
   }
 }
