+++ apps/router/src/languages/ko.json
@@ -32,7 +32,8 @@
     "connect": "연결하다",
     "password": "비밀번호",
     "federation-name": "연합 이름",
-    "guardian-name": "보호자 이름"
+    "guardian-name": "보호자 이름",
+    "guardians": "가디언즈"
   },
   "notConfigured": {
     "title": "구성되지 않음",
@@ -417,7 +418,16 @@
     "to-address": "주소 지정하기",
     "txid": "거래 ID",
     "claimed-note": "주장된 메모",
-    "claimed-amount": "청구 금액"
+    "claimed-amount": "청구 금액",
+    "balances-summary": "잔액 요약",
+    "lightning-payments": "라이트닝 결제",
+    "lightning-payments-description": "연합을 통해 라이트닝 네트워크 결제를 보내고 받으세요.",
+    "onchain-description": "연합에 비트코인을 입금하고 출금하세요.",
+    "gateway-onchain-manage": "게이트웨이의 온체인 지갑에서 비트코인을 관리하세요.",
+    "federation-balances": "연합 잔액",
+    "peg-in": "페그-인",
+    "peg-out": "페그 아웃",
+    "gateway-on-chain-wallet": "게이트웨이 온체인 지갑"
   },
   "wallet-modal": {
     "title": "지갑 작업",
@@ -443,7 +453,11 @@
       "peg-out-success": "Peg Out TX 전송 완료!",
       "create-ecash": "Ecash 노트 생성하기",
       "ecash-created": "당신의 {{federationName}} 잔액에서 {{amount}} ecash 노트를 인출했습니다. 수령인에게 보내서 청구하세요.",
-      "ecash-error": "ecash 메모 생성 오류: {{error}}"
+      "ecash-error": "ecash 메모 생성 오류: {{error}}",
+      "lightning-invoice": "라이트닝 인보이스",
+      "lightning-invoice-placeholder": "lnbc...",
+      "pay-invoice": "라이트닝 청구서 지불",
+      "peg-out-error": "온체인에서 핀 고정 실패"
     }
   },
   "federation-card": {
@@ -556,5 +570,14 @@
     "sentence-one": "주소로 입금하거나 QR 코드를 스캔하세요.",
     "transaction-id": "거래 ID",
     "create-pegin-address": "페그인 주소 생성"
+  },
+  "router": {
+    "title": "아직 연결된 서비스가 없습니다.",
+    "services-title": "페디민트 연합은 두 가지 유형의 서비스로 구성됩니다:",
+    "services-description": "Fedimint 프로토콜 운영, 자금 관리 및 eCash 노트의 발행 및 상환 관리를 담당합니다. 그들은 분산 합의를 사용하여 연합을 보호합니다.",
+    "gateways": "게이트웨이 (라이트닝 서비스 제공자):",
+    "gateways-description": "페디민트와 라이트닝 네트워크 간의 브리지로, 사용자가 라이트닝 결제를 송금하고 수신할 수 있게 해줍니다. 이는 연합 또는 독립 제공자가 운영할 수 있습니다.",
+    "learn-more": "여기에서 Fedimint 서비스를 설정하는 방법에 대해 자세히 알아볼 수 있습니다:",
+    "learn-more-link": "Fedimint 서비스를 설정하는 방법을 배우세요."
   }
 }
