+++ apps/router/src/languages/pt.json
@@ -32,7 +32,8 @@
     "connect": "Conectar",
     "password": "Senha",
     "federation-name": "Nome da Federação",
-    "guardian-name": "Nome do Guardião"
+    "guardian-name": "Nome do Guardião",
+    "guardians": "Guardians"
   },
   "notConfigured": {
     "title": "Não Configurado",
@@ -417,7 +418,16 @@
     "to-address": "Para Endereço",
     "txid": "ID da Transação",
     "claimed-note": "Nota Reivindicada",
-    "claimed-amount": "Valor Reivindicado"
+    "claimed-amount": "Valor Reivindicado",
+    "balances-summary": "Resumo de Saldos",
+    "lightning-payments": "Pagamentos Lightning",
+    "lightning-payments-description": "Envie e receba pagamentos da Lightning Network através da federação",
+    "onchain-description": "Deposite e retire bitcoin da/para a federação",
+    "gateway-onchain-manage": "Gerencie bitcoin na carteira onchain do gateway",
+    "federation-balances": "Saldos da Federação",
+    "peg-in": "Peg-In",
+    "peg-out": "Peg-Out",
+    "gateway-on-chain-wallet": "Carteira On-Chain Gateway"
   },
   "wallet-modal": {
     "title": "Ações da Carteira",
@@ -443,7 +453,11 @@
       "peg-out-success": "Peg Out TX Enviado!",
       "create-ecash": "Criar Nota Ecash",
       "ecash-created": "Retirou uma nota de ecash de {{amount}} do seu saldo {{federationName}}. Envie-a para o destinatário para reivindicar.",
-      "ecash-error": "Erro ao criar nota ecash: {{error}}"
+      "ecash-error": "Erro ao criar nota ecash: {{error}}",
+      "lightning-invoice": "Fatura Relâmpago",
+      "lightning-invoice-placeholder": "lnbc...",
+      "pay-invoice": "Pagar Fatura Lightning",
+      "peg-out-error": "Falha ao fixar na cadeia."
     }
   },
   "federation-card": {
@@ -553,5 +567,14 @@
     "sentence-one": "Deposite no endereço ou escaneie o código QR",
     "transaction-id": "ID da Transação",
     "create-pegin-address": "Criar Endereço Peg-in"
+  },
+  "router": {
+    "title": "Nenhum serviço conectado ainda.",
+    "services-title": "Uma federação Fedimint consiste em dois tipos de serviços:",
+    "services-description": "Responsável por executar o protocolo Fedimint, custodiar fundos e gerenciar a emissão e resgate de notas eCash. Eles utilizam consenso distribuído para garantir a segurança da federação.",
+    "gateways": "Gateways (Provedores de Serviço Lightning):",
+    "gateways-description": "Ponte entre o Fedimint e a Lightning Network, permitindo que os usuários enviem e recebam pagamentos Lightning. Podem ser operados pela federação ou por provedores independentes.",
+    "learn-more": "Você pode aprender mais sobre como configurar os serviços do Fedimint aqui:",
+    "learn-more-link": "Aprenda como configurar os serviços Fedimint."
   }
 }
