+++ packages/types/src/gateway.ts
@@ -45,9 +45,9 @@ export interface FederationInfo {
 export interface CreateBolt11InvoiceV2Payload {
   federation_id: string;
   contract: IncomingContract;
-  invoice_amount: number;
-  description: Bolt11InvoiceDescription;
-  expiry_time: number;
+  amount_msats: number;
+  description: string;
+  expiry_secs: number;
 }
 
 export interface IncomingContract {
@@ -241,4 +241,18 @@ export interface GatewayBalances {
 export interface FederationBalanceInfo {
   federation_id: string;
   ecash_balance_msats: number;
+}
+
+export interface WithdrawResponse {
+  txid: string;
+  fees: PegOutFees;
+}
+
+export interface PegOutFees {
+  fee_rate: FeeRate;
+  total_weight: number;
+}
+
+export interface FeeRate {
+  sats_per_kvb: number;
 }
